+++ misskey-websocket/src/broker.rs
@@ -10,7 +10,7 @@ use crate::model::outgoing::OutgoingMessage;
 use async_std::task;
 #[cfg(feature = "async-std-runtime")]
 use async_std::task::sleep;
-use async_tungstenite::tungstenite::http::{header::HeaderMap, Uri};
+use async_tungstenite::tungstenite::http::HeaderMap;
 use async_tungstenite::tungstenite::Error as WsError;
 use futures_util::stream::StreamExt;
 use log::{info, warn};
@@ -18,6 +18,7 @@ use log::{info, warn};
 use tokio::task;
 #[cfg(feature = "tokio-runtime")]
 use tokio::time::sleep;
+use url::Url;
 
 pub mod channel;
 pub mod handler;
@@ -32,7 +33,7 @@ pub(crate) struct Broker {
     broker_rx: ControlReceiver,
     handler: Handler,
     reconnect: ReconnectConfig,
-    uri: Uri,
+    url: Url,
     additional_headers: HeaderMap,
 }
 
@@ -180,9 +181,9 @@ impl Default for ReconnectConfig {
 
 impl Broker {
     pub async fn spawn(
-        uri: Uri,
-        reconnect: ReconnectConfig,
+        url: Url,
         additional_headers: HeaderMap,
+        reconnect: ReconnectConfig,
     ) -> Result<(ControlSender, SharedBrokerState)> {
         let state = SharedBrokerState::working();
         let shared_state = SharedBrokerState::clone(&state);
@@ -191,7 +192,7 @@ impl Broker {
 
         task::spawn(async move {
             let mut broker = Broker {
-                uri,
+                url,
                 additional_headers,
                 broker_rx,
                 reconnect,
@@ -265,7 +266,7 @@ impl Broker {
         use futures_util::future::{self, Either};
 
         let (mut websocket_tx, mut websocket_rx) =
-            match connect_websocket(self.uri.clone(), self.additional_headers.clone()).await {
+            match connect_websocket(self.url.clone(), self.additional_headers.clone()).await {
                 Ok(x) => x,
                 Err(error) => {
                     // retain `remaining_message` because we've not sent it yet