+++ misskey-websocket/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+- Accept additional headers for `WebSocketClient`
+
 ### Changed
 
 - Use tokio 1.0 and async-tungstenite 0.13